+++ hutool-core/src/main/java/com/xiaoleilu/hutool/util/BeanUtil.java
@@ -15,6 +15,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 
+import com.xiaoleilu.hutool.bean.BeanInfoCache;
 import com.xiaoleilu.hutool.collection.CaseInsensitiveMap;
 import com.xiaoleilu.hutool.convert.Convert;
 import com.xiaoleilu.hutool.exceptions.UtilException;
@@ -60,6 +61,7 @@ public final class BeanUtil {
 		return PropertyEditorManager.findEditor(type);
 	}
 
+	//--------------------------------------------------------------------------------------------------------- PropertyDescriptor
 	/**
 	 * 获得Bean字段描述数组
 	 * 
@@ -72,26 +74,42 @@ public final class BeanUtil {
 	}
 	
 	/**
-	 * 获得字段名和字段描述Map
+	 * 获得字段名和字段描述Map，获得的结果会缓存在 {@link BeanInfoCache}中
 	 * @param clazz Bean类
 	 * @param ignoreCase 是否忽略大小写
 	 * @return 字段名和字段描述Map
 	 * @throws IntrospectionException 获取属性异常
 	 */
-	public static Map<String, PropertyDescriptor> getFieldNamePropertyDescriptorMap(Class<?> clazz, boolean ignoreCase) throws IntrospectionException{
+	public static Map<String, PropertyDescriptor> getPropertyDescriptorMap(Class<?> clazz, boolean ignoreCase) throws IntrospectionException{
+		Map<String, PropertyDescriptor> map = BeanInfoCache.INSTANCE.getPropertyDescriptorMap(clazz, ignoreCase);
+		if(null == map){
+			map = internalGetPropertyDescriptorMap(clazz, ignoreCase);
+			BeanInfoCache.INSTANCE.putPropertyDescriptorMap(clazz, map, ignoreCase);
+		}
+		return map;
+	}
+	
+	/**
+	 * 获得字段名和字段描述Map。内部使用，直接获取Bean类的PropertyDescriptor
+	 * @param clazz Bean类
+	 * @param ignoreCase 是否忽略大小写
+	 * @return 字段名和字段描述Map
+	 * @throws IntrospectionException 获取属性异常
+	 */
+	private static Map<String, PropertyDescriptor> internalGetPropertyDescriptorMap(Class<?> clazz, boolean ignoreCase) throws IntrospectionException{
 		final PropertyDescriptor[] propertyDescriptors = getPropertyDescriptors(clazz);
-		Map<String, PropertyDescriptor> map = ignoreCase ?
-				new CaseInsensitiveMap<String, PropertyDescriptor>(propertyDescriptors.length)
-				: new HashMap<String, PropertyDescriptor>((int) (propertyDescriptors.length / 0.75));
+		final Map<String, PropertyDescriptor> map = ignoreCase ?
+				new CaseInsensitiveMap<String, PropertyDescriptor>(propertyDescriptors.length, 1)
+				: new HashMap<String, PropertyDescriptor>((int) (propertyDescriptors.length), 1);
 				
 		for (PropertyDescriptor propertyDescriptor : propertyDescriptors) {
 			map.put(propertyDescriptor.getName(), propertyDescriptor);
 		}
 		return map;
 	}
-
+	
 	/**
-	 * 获得Bean类属性描述
+	 * 获得Bean类属性描述，大小写敏感
 	 * 
 	 * @param clazz Bean类
 	 * @param fieldName 字段名
@@ -99,13 +117,21 @@ public final class BeanUtil {
 	 * @throws IntrospectionException 获取属性异常
 	 */
 	public static PropertyDescriptor getPropertyDescriptor(Class<?> clazz, final String fieldName) throws IntrospectionException {
-		PropertyDescriptor[] propertyDescriptors = getPropertyDescriptors(clazz);
-		for (PropertyDescriptor propertyDescriptor : propertyDescriptors) {
-			if (ObjectUtil.equal(fieldName, propertyDescriptor.getName())) {
-				return propertyDescriptor;
-			}
-		}
-		return null;
+		return getPropertyDescriptor(clazz, fieldName, false);
+	}
+	
+	/**
+	 * 获得Bean类属性描述
+	 * 
+	 * @param clazz Bean类
+	 * @param fieldName 字段名
+	 * @param ignoreCase 是否忽略大小写
+	 * @return PropertyDescriptor
+	 * @throws IntrospectionException 获取属性异常
+	 */
+	public static PropertyDescriptor getPropertyDescriptor(Class<?> clazz, final String fieldName, boolean ignoreCase) throws IntrospectionException {
+		final Map<String, PropertyDescriptor> map = getPropertyDescriptorMap(clazz, ignoreCase);
+		return (null == map) ? null : map.get(fieldName);
 	}
 	
 	/**
@@ -499,7 +525,7 @@ public final class BeanUtil {
 		
 		final Map<String, PropertyDescriptor> sourcePdMap;
 		try {
-			sourcePdMap = getFieldNamePropertyDescriptorMap(source.getClass(), ignoreCase);
+			sourcePdMap = getPropertyDescriptorMap(source.getClass(), ignoreCase);
 		} catch (IntrospectionException e) {
 			throw new UtilException(e);
 		}